+++ test/fixtures/cjs/module_a_part1.js
@@ -1,0 +1,14 @@
+/**
+ * @module Cache
+ *
+ * Something.
+ */
+var Cache = {
+};
+
+/**
+ * Add an item to the cache.
+ */
+Cache.add = function() {};
+
+module.exports = Cache;