+++ test/fixtures/cjs/module_a_part2.js
@@ -1,0 +1,7 @@
+/**
+ * @memberOf Cache
+ *
+ * Remove an item from the cache.
+ */
+function remove() {
+}